feat(services): Implement several core features for interation and routing chore: Rename index.js to server.js build: Add nodemon and amend package.json docs: Add JSDoc convention commenting
The following API functionalities were implemented and routed:
- getTickerInfo
- getTickerInfoWithClose
- getTickerScore
- getAllTickersInfo (Takes an optional prices boolean param)

These can be tested on routes:
- '/': getAllTickersInfo
- '/:ticker': getTickerInfoWithClose
- '/:ticker/score': getTickerScore

The rationale behind implementing JSDoc commenting for functions on such
a small project is to demonstrate initiative in thinking regarding
scaling, and to treat this as one would production code. JSDoc comments
allow for easy documentation generation across a large codebase,
implementing this from the start is the easiest way to avoid backloading
the documentation to some (likely painful) point in the future.